factored out more application specifics. added a static app builder.

+++ src/build/makers/android.js
@@ -8,7 +8,6 @@ var shell        = require('shelljs'),
     fs           = require('fs');
 
 var android_lib = libraries.paths['cordova-android'];
-var mobile_spec = libraries.output.test;
 var create = path.join(android_lib, 'bin', 'create');
 
 module.exports = function(output, sha, devices, entry_point, callback) {
@@ -39,11 +38,13 @@ module.exports = function(output, sha, devices, entry_point, callback) {
                         if (!fs.existsSync(output)) {
                             throw new Error('./bin/create must have failed as output path does not exist.');
                         }
-                        shell.cp('-Rf', path.join(mobile_spec, '*'), path.join(output, 'assets', 'www'));
+                        shell.cp('-Rf', path.join(libraries.output.test, '*'), path.join(output, 'assets', 'www'));
                         
                         // add the sha to the junit reporter
                         var tempJasmine = path.join(output, 'assets', 'www', 'jasmine-jsreporter.js');
-                        fs.writeFileSync(tempJasmine, "var library_sha = '" + sha + "';\n" + fs.readFileSync(tempJasmine, 'utf-8'), 'utf-8');
+                        if (fs.existsSync(tempJasmine)) {
+                            fs.writeFileSync(tempJasmine, "var library_sha = '" + sha + "';\n" + fs.readFileSync(tempJasmine, 'utf-8'), 'utf-8');
+                        }
 
                         // modify start page
                         // 1. old cordova-android: modify the .java file
@@ -54,9 +55,11 @@ module.exports = function(output, sha, devices, entry_point, callback) {
                         fs.writeFileSync(configFile, fs.readFileSync(configFile, 'utf-8').replace(/<content\s*src=".*"/gi, '<content src="' +entry_point + '"'), 'utf-8');
                         
                         // look at which cordova-<v>.js current lib uses
-                        var version = fs.readFileSync(path.join(android_lib, 'VERSION'), 'utf-8').replace(/\r?\n/,'');
                         var cordovajs = path.join(output, 'assets', 'www', 'cordova.js');
-                        fs.writeFileSync(cordovajs, fs.readFileSync(cordovajs, 'utf-8').replace(/var VERSION='.*';/, "var VERSION='" + version + "';"), 'utf-8');
+                        if (fs.existsSync(cordovajs)) {
+                            var version = fs.readFileSync(path.join(android_lib, 'VERSION'), 'utf-8').replace(/\r?\n/,'');
+                            fs.writeFileSync(cordovajs, fs.readFileSync(cordovajs, 'utf-8').replace(/var VERSION='.*';/, "var VERSION='" + version + "';"), 'utf-8');
+                        }
                     } catch (e) {
                         error_writer('android', sha, 'Exception thrown modifying Android mobile spec application.', e.message);
                         callback(true);
