+++ src/build/makers/blackberry.js
@@ -8,7 +8,6 @@ var shell        = require('shelljs'),
     fs           = require('fs');
 
 var blackberry_lib = libraries.paths['cordova-blackberry'];
-var mobile_spec = libraries.output.test;
 var create = path.join(blackberry_lib, 'bin', 'create');
 
 module.exports = function(output, sha, devices, entry_point, callback) {
@@ -34,11 +33,13 @@ module.exports = function(output, sha, devices, entry_point, callback) {
                     try {
                         // copy over mobile spec modified html assets
                         log('Modifying Cordova application.');
-                        shell.cp('-Rf', path.join(mobile_spec, '*'), path.join(output, 'www'));
+                        shell.cp('-Rf', path.join(libraries.output.test, '*'), path.join(output, 'www'));
 
                         // drop the BlackBerry library SHA into the junit reporter
                         var tempJasmine = path.join(output, 'www', 'jasmine-jsreporter.js');
-                        fs.writeFileSync(tempJasmine, "var library_sha = '" + sha + "';\n" + fs.readFileSync(tempJasmine, 'utf-8'), 'utf-8');
+                        if (fs.existsSync(tempJasmine)) {
+                            fs.writeFileSync(tempJasmine, "var library_sha = '" + sha + "';\n" + fs.readFileSync(tempJasmine, 'utf-8'), 'utf-8');
+                        }
 
                         // TODO: make sure we are using blackberry lib's version in mobile-spec's cordoa.js
